requirements.txt
asgiref==3.11.0       
django==6.0
django-jazzmin==3.0.1
django-nested-admin==4.1.6
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8

sqlparse==0.5.4
tzdata==2025.3
whitenoise==6.11.0